# openapi.yaml
openapi: 3.0.3
info:
  title: Booksy API
  version: 1.0.0
...
paths:
...
  /purchase/preview:
    post:
      summary: Preview the price, discounts, and loyalty points for a purchase
      tags:
        - purchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchasePreviewRequest'
      responses:
        '200':
          description: Purchase preview calculated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchasePreviewResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiError'
  /purchase:
    post:
      summary: Purchase one or more books
      tags:
        - purchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseRequest'
      responses:
        '200':
          description: Purchase successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseResponse'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
...
    PurchaseRequest:
      type: object
      properties:
        customerId:
          type: integer
        bookIds:
          type: array
          items:
            type: integer
        agreedTotalPrice:
          type: number
          format: double
        loyaltyRewardApplied:
          type: boolean
      required: [customerId, bookIds, agreedTotalPrice, loyaltyRewardApplied]
    PurchasePreviewRequest:
      type: object
      properties:
        customerId:
          type: integer
        bookIds:
          type: array
          items:
            type: integer
      required: [ customerId, bookIds ]
    PurchasePreviewResponse:
      type: object
      properties:
        customerId:
          type: integer
        totalPrice:
          type: number
          format: double
        discountApplied:
          type: number
          format: double
        loyaltyPointsEarned:
          type: integer
        loyaltyRewardApplied:
          type: boolean
      required: [customerId, totalPrice, discountApplied, loyaltyPointsEarned, loyaltyRewardApplied]
    PurchaseResponse:
      type: object
      properties:
        purchaseId:
          type: integer
        totalPrice:
          type: number
          format: double
        loyaltyPointsEarned:
          type: integer
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        status:
          type: integer
        timestamp:
          type: string
        fieldErrors:
          type: array
          items:
            type: string